+++ apps/web/src/pages/dashboard/campaigns/[id]/campaign-actions.tsx
@@ -87,7 +87,8 @@ export const CampaignActions = () => {
   const content = form.watch("content") || ""
 
   const finalContent = useMemo(() => {
-    const template = campaignQuery.data?.campaign.Template?.content || ""
+    const template =
+      campaignQuery.data?.campaign.Template?.content || "{{content}}"
     const final = template.replace(/{{content}}/g, content)
     return final
   }, [campaignQuery.data?.campaign.Template?.content, content])